created utils and updated index
